Added x64 support on Linux

+++ core/include/core/inline/Rectangle.inl
@@ -22,7 +22,7 @@
 
 Int32 Rectangle::bottom() const
 {
-	return y + height;
+	return y + static_cast<Int32>(height);
 }
 
 Int32 Rectangle::left() const
@@ -32,7 +32,7 @@ Int32 Rectangle::left() const
 
 Int32 Rectangle::right() const
 {
-	return x + width;
+	return x + static_cast<Int32>(width);
 }
 
 Int32 Rectangle::top() const
